Addition of GSON jar...
I spoke with Professor Waters and he said it would be okay for use to
use GSON as our JSON parser, I have added it to the library folder and
corrected the build path to include it. I am going to be working on
including this in all of the files to pass information between intents.